refactor(server): extract image helpers → src/server/images.js
Pure move, zero behavior change. Pulls the four image-generation helpers
out of server.js into src/server/images.js:
  - buildImagePrompt / generateHeroImage         — article hero, 16:9 (5/7 refs)
  - buildSocialImagePrompt / generateSocialImage  — social post, 1:1 (2/2 refs)
plus the shared HERO_IMAGE_NEGATIVE_PROMPT const (kept module-private —
only the two fal.ai generators use it).

Why: continues the server.js dismemberment. These four are a cohesive
media-generation cluster. Their only non-global dependency is the shared
`anthropic` client, which already lives in src/server/llm.js — so the
module imports it from there. Otherwise just fetch + process.env
(FAL_API_KEY).

What:
- New src/server/images.js exports the four functions (moved verbatim),
  imports `anthropic` from ./llm.js. HERO_IMAGE_NEGATIVE_PROMPT private.
- server.js imports the four; inline defs (two separate blocks, hero pair
  ~L867 + social pair ~L7355) replaced with breadcrumb comments.

Test plan:
- node --check server.js -> OK
- npm run lint (no-undef gate) -> clean (confirms all four imported, no
  orphaned HERO_IMAGE_NEGATIVE_PROMPT reference left behind)
- route-inventory guard -> PASS (snapshot unchanged)
- vitest -> 65/65 pass (+5 new tests: buildImagePrompt text/fallback/
  brand-visual-branch via a mocked anthropic client; buildSocialImagePrompt
  text + fallback). The two fal.ai generators are pure network wrappers —
  exercised via the live publish path, not unit-tested.

Rollback: revert; pure copy, restores inline defs exactly.
